requirements.txt
langchain>=0.3.12
langchain-openai>=0.2.12
langchainhub>=0.1.21
langchain-community>=0.3.12
python-dotenv>=1.0.1
langgraph>=0.2.59
jinja2>=3.1.4